moved more functions into page.modules.js
